@@ -2,9 +2,9 @@ package net.aluraflix.exception;
 
 import io.quarkus.logging.Log;
 
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
 
 public class EntityNotFoundException extends RuntimeException{
 
