@@ -3,8 +3,8 @@ package net.aluraflix.model.video;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.hibernate.validator.constraints.URL;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 @Schema(name = "Video", description = "Movie representation")
 public record VideoForm(@Schema(example = "Title", required = true) @NotBlank String titulo,
