@@ -9,9 +9,9 @@ import net.aluraflix.model.category.Category;
 import net.aluraflix.model.category.CategoryRepository;
 import net.aluraflix.model.video.*;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.transaction.Transactional;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
