@@ -1,6 +1,6 @@
  const URL_BACKEND_TOP = window.location.hostname.includes('localhost')
-   ? 'http://localhost:8080'
-   : 'http://aluraflix-wmime.net:8080';
+   ? 'http://localhost'
+   : 'http://localhost:8080';
 
 export default {
   URL_BACKEND_TOP,
