@@ -1,13 +1,13 @@
 package net.aluraflix.service;
 
+import io.quarkus.elytron.security.common.BcryptUtil;
 import io.quarkus.logging.Log;
+import jakarta.enterprise.context.ApplicationScoped;
 import net.aluraflix.exception.EmailAlreadyRegisteredException;
 import net.aluraflix.exception.InvalidCredentialsException;
 import net.aluraflix.model.client.Client;
 import net.aluraflix.model.client.ClientForm;
-import org.apache.sshd.common.config.keys.loader.openssh.kdf.BCrypt;
 
-import javax.enterprise.context.ApplicationScoped;
 import java.util.Optional;
 
 @ApplicationScoped
@@ -19,7 +19,7 @@ public class AuthService {
             throw new InvalidCredentialsException("Invalid username");
         }
         Client client = clientOptional.get();
-        if (!BCrypt.checkpw(clientForm.password(), client.getPassword())) {
+        if (!BcryptUtil.matches(clientForm.password(), client.getPassword())) {
             throw new InvalidCredentialsException("Invalid password");
         }
         Log.info("Successful login.");
